+++ tests/PointSettingsTest.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace InfluxDB2Test;
+
+
+use InfluxDB2\Client;
+use InfluxDB2\Model\WritePrecision;
+use PHPUnit\Framework\TestCase;
+
+class PointSettingsTest extends TestCase
+{
+    private const ID_TAG = "132-987-655";
+    private const CUSTOMER_TAG = "California Miner";
+
+    /** @var Client */
+    private $client;
+
+    public function setUp()
+    {
+        $this->client = new Client([
+            "url" => "http://localhost:9999",
+            "token" => "my-token",
+            "bucket" => "my-bucket",
+            "precision" => WritePrecision::NS,
+            "org" => "my-org"
+        ]);
+    }
+
+    public function testPointSettings()
+    {
+        $writeApi = $this->client->createWriteApi(null, array('id' => PointSettingsTest::ID_TAG,
+            'customer' => PointSettingsTest::CUSTOMER_TAG));
+
+        $defaultTags = $writeApi->pointSettings->getDefaultTags();
+
+        $this->assertEquals(PointSettingsTest::ID_TAG, $defaultTags['id']);
+        $this->assertEquals(PointSettingsTest::CUSTOMER_TAG, $defaultTags['customer']);
+    }
+
+    public function testPointSettingsWithAdd()
+    {
+        putenv("data_center=LA");
+
+        $writeApi = $this->client->createWriteApi();
+
+        $writeApi->pointSettings->addDefaultTag('id', PointSettingsTest::ID_TAG);
+        $writeApi->pointSettings->addDefaultTag('customer', PointSettingsTest::CUSTOMER_TAG);
+        $writeApi->pointSettings->addDefaultTag('data_center', '${env.data_center}');
+
+        $defaultTags = $writeApi->pointSettings->getDefaultTags();
+
+        $this->assertEquals(PointSettingsTest::ID_TAG, $defaultTags['id']);
+        $this->assertEquals(PointSettingsTest::CUSTOMER_TAG, $defaultTags['customer']);
+        $this->assertEquals("LA", $defaultTags['data_center']);
+    }
+}